update react-select docs structure
