--- docs/advanced/index.js
@@ -1,11 +1,0 @@
-// @flow
-
-import md from '../markdown/renderer';
-
-export default function Advanced() {
-  return md`
-    # Advanced
-
-    Content coming soon...
-  `;
-}
