--- docs/api/examples/index.js
@@ -1,1 +1,0 @@
-export { default as PropTypes } from './PropTypes';
