--- docs/async/index.js
@@ -1,11 +1,0 @@
-// @flow
-
-import md from '../markdown/renderer';
-
-export default function Async() {
-  return md`
-    # Async
-
-    Content coming soon...
-  `;
-}
