--- docs/creatable/index.js
@@ -1,11 +1,0 @@
-// @flow
-
-import md from '../markdown/renderer';
-
-export default function Creatable() {
-  return md`
-    # Creatable
-
-    Content coming soon...
-  `;
-}
