--- docs/home/examples/AsyncCallbacks.js
@@ -1,41 +1,0 @@
-import React, { Component } from 'react';
-
-import AsyncSelect from '../../../src/Async';
-import { colourOptions } from '../../data';
-
-type State = {
-  inputValue: string,
-};
-
-const filterColors = (inputValue: string) =>
-  colourOptions.filter(i =>
-    i.label.toLowerCase().includes(inputValue.toLowerCase())
-  );
-
-const loadOptions = (inputValue, callback) => {
-  setTimeout(() => {
-    callback(filterColors(inputValue));
-  }, 1000);
-};
-
-export default class WithCallbacks extends Component<*, State> {
-  state = { inputValue: '' };
-  handleInputChange = (newValue: string) => {
-    const inputValue = newValue.replace(/\W/g, '');
-    this.setState({ inputValue });
-    return inputValue;
-  };
-  render() {
-    return (
-      <div>
-        <pre>inputValue: "{this.state.inputValue}"</pre>
-        <AsyncSelect
-          cacheOptions
-          loadOptions={loadOptions}
-          defaultOptions
-          onInputChange={this.handleInputChange}
-        />
-      </div>
-    );
-  }
-}
