--- docs/home/examples/BasicMulti.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import Select from '../../../src';
-import { colourOptions } from '../../data';
-
-export default () => (
-  <Select
-    defaultValue={[colourOptions[2], colourOptions[3]]}
-    isMulti
-    name="colors"
-    options={colourOptions}
-  />
-);
