+++ docs/pages/advanced/index.js
@@ -1,0 +1,10 @@
+// @flow
+import md from '../../markdown/renderer';
+
+export default function Advanced() {
+  return md`
+    # Advanced
+
+    Content coming soon...
+  `;
+}
