+++ docs/pages/api/index.js
@@ -1,0 +1,53 @@
+// @flow
+import md from '../../markdown/renderer';
+import { Selects } from '../../PropTypes';
+
+export default function Api() {
+  return md`
+    # API
+
+    ## Methods
+
+    ### \`focus()\`
+
+    Focused the internal control input.
+
+    ## Prop Types
+
+    ### Internal Types
+
+    You'll see these in the public props below:
+
+    ~~~js
+    type OptionType = { [string]: any }
+    type OptionsType = Array<OptionType>
+
+    type GroupType = {
+      [string]: any, // group label
+      options: OptionsType,
+    }
+
+    type ValueType = OptionType | OptionsType | null | void
+
+    type CommonProps = {
+      clearValue: () => void,
+      getStyles: (string, any) => {},
+      getValue: () => ValueType,
+      hasValue: boolean,
+      isMulti: boolean,
+      options: OptionsType,
+      selectOption: OptionType => void,
+      selectProps: any,
+      setValue: (ValueType, ActionTypes) => void,
+    }
+
+    // passed as the second argument to \`onChange\`
+    type ActionTypes = | 'clear' | 'create-option' | 'deselect-option' | 'pop-value' | 'remove-value' | 'select-option' | 'set-value'
+    ~~~
+
+    ### Public Props
+
+    ${Selects.Base}
+
+  `;
+}
