+++ docs/pages/async/index.js
@@ -1,0 +1,11 @@
+// @flow
+
+import md from '../../markdown/renderer';
+
+export default function Async() {
+  return md`
+    # Async
+
+    Content coming soon...
+  `;
+}
