+++ docs/pages/creatable/index.js
@@ -1,0 +1,11 @@
+// @flow
+
+import md from '../../markdown/renderer';
+
+export default function Creatable() {
+  return md`
+    # Creatable
+
+    Content coming soon...
+  `;
+}
